SDK Nightly build 2026-04-16
List of changes from previous build

## mcu_plus_sdk
Previous: d7e166b58d93 -> Current: a99d9f54be94

a99d9f54be9 am261x: iar: Fix issue with the TIREX
eed4c5d72a0 am26xx: docs: Add documentation for MII mode

## security_common
Previous: 192441761fa6 -> Current: a4844abf24db

a4844ab f29h85x: misra: MISRA.ETYPE.ASSIGN.2012

+++ source/security/security_common/drivers/crypto/pke/tifs_pke4_random.c
@@ -123,7 +123,7 @@ int32_t cri_pke_get_pseudo_random(void *buf, size_t len, int32_t slot, uint32_t 
 			if (slot == CRI_PKE_NO_SLOT) {
 				(void)memcpy(buf, pke_addr(0, NULL, bits), len);
 			} else {
-				(void)memcpy(buf, pke_addr(slot, NULL, bits), len);
+				(void)memcpy(buf, pke_addr((uint32_t)slot, NULL, bits), len);
 			}
 		}
 	}
